Policies: Make religion policy more fair
Currently it was only possible to use this policy if priety was completed. So people who uses Order couldn't use this one :-(

Now it is available after discovering Theology and Printing press.

In the future it should be available after completing (maybe four) policies. See https://github.com/yairm210/Unciv/issues/3242#issuecomment-1514617790 for the Feature Request.